requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
gpiod==2.3.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pupil_apriltags==1.0.4.post11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
